+++ src/main/kotlin/com/autonomousapps/advice/PluginAdvice.kt
@@ -11,12 +11,21 @@ data class PluginAdvice(
   companion object {
     fun redundantJavaLibrary() = PluginAdvice(
       redundantPlugin = "java-library",
-      reason = "this project has both java-library and org.jetbrains.kotlin.jvm applied, which is redundant. You can remove java-library"
+      reason = "this project has both java-library and org.jetbrains.kotlin.jvm applied, which " +
+        "is redundant. You can remove java-library"
     )
 
     fun redundantKotlinJvm() = PluginAdvice(
       redundantPlugin = "org.jetbrains.kotlin.jvm",
-      reason = "this project has both java-library and org.jetbrains.kotlin.jvm applied, which is redundant. You can remove org.jetbrains.kotlin.jvm"
+      reason = "this project has both java-library and org.jetbrains.kotlin.jvm applied, which " +
+        "is redundant. You can remove org.jetbrains.kotlin.jvm"
+    )
+
+    fun redundantKapt() = PluginAdvice(
+      redundantPlugin = "kotlin-kapt",
+      reason = "this project has the kotlin-kapt (org.jetbrains.kotlin.kapt) plugin applied, but " +
+        "no annotation processors (or no used annotation processors), which is redundant. You " +
+        "can remove it"
     )
   }