+++ src/main/kotlin/com/autonomousapps/internal/outputDirectories.kt
@@ -72,8 +72,10 @@ fun getAdvicePrettyPath(variantName: String) =
 fun getAdviceConsolePath(variantName: String) =
   "${getVariantDirectory(variantName)}/advice.txt"
 
-// Redundant plugin tasks (nb: no variants involved)
-fun getPluginAdvicePath() = "$ROOT_DIR/intermediates/advice-plugin.json"
+// Redundant plugin tasks (nb: variants not necessarily involved)
+fun getPluginJvmAdvicePath() = "$ROOT_DIR/advice-plugin-jvm.json"
+fun getPluginKaptAdvicePath(variantName: String) =
+  "${getVariantDirectory(variantName)}/advice-plugin-kapt.json"
 
 // Root project aggregate reports. No need for variant-specific directory
 fun getMisusedDependenciesAggregatePath() = "$ROOT_DIR/intermediates/misused-dependencies.json"